requirements.txt
opencv-python>=4.8.0,<5.0.0
ultralytics>=8.0.0,<9.0.0
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
